Trim whitespace around each history entry, before removing dups.
Especially useful when tab-completion in the shell's input reading adds an extra
trailing space to many commands, and also useful for other situations that
somehow have surrounding whitespace.  Now, entries (shell commands) are compared
for equality by ignoring such surrounding whitespace, which prevents having
duplicates that only differed in such.  Now, the combined entries that are
output do not include such surrounding whitespace, which normalizes them.

Whitespace that is not leading or trailing is still significant, because it'd be
too impractical to identify which is and is not safe to normalize.